# src/lib/sdk/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Scalius Commerce API",
    "version": "1.0.0",
    "description": "API for the Scalius Commerce platform. This specification provides all the necessary details to build a complete e-commerce storefront, from displaying products and categories to processing orders and managing customer data.",
    "contact": {
      "name": "Scalius Support",
      "url": "https://scalius.com/support"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A short-lived JSON Web Token obtained from the /auth/token endpoint."
      },
      "apiToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token",
        "description": "A static, pre-shared token for machine-to-machine authentication to obtain a JWT."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error Type"
          },
          "message": {
            "type": "string",
            "example": "A detailed error message."
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/categories/{slug}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get category by slug",
        "description": "Returns detailed information about a specific category",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Slug of the category to retrieve",
            "schema": {
              "type": "string",
              "example": "tech-gadgets"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "category": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "cat_fj1ZspPwVuHyQVCGDC0jf"
                        },
                        "name": {
                          "type": "string",
                          "example": "Tech Gadgets"
                        },
                        "slug": {
                          "type": "string",
                          "example": "tech-gadgets"
                        },
                        "description": {
                          "type": "string",
                          "example": "🔌 From smart tools to must-have devices — tech that simplifies & empowers everyday life."
                        },
                        "imageUrl": {
                          "type": "string",
                          "nullable": true,
                          "example": null
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "+057378-04-03T14:00:00.000Z"
                        },
                        "metaTitle": {
                          "type": "string",
                          "nullable": true,
                          "example": "Cool Tech Gadgets Online – Smart Tools"
                        },
                        "metaDescription": {
                          "type": "string",
                          "nullable": true,
                          "example": "Shop trending tech gadgets that make life easier & smarter. Functional, innovative & cool."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/navigation": {
      "get": {
        "tags": [
          "Navigation"
        ],
        "summary": "Get navigation menus",
        "description": "Returns navigation menu structure for header and/or footer",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Type of navigation to retrieve",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "header",
                "footer",
                "all"
              ],
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved navigation data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "navigation": {
                      "type": "object",
                      "properties": {
                        "header": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "title": {
                                "type": "string",
                                "example": "Home"
                              },
                              "href": {
                                "type": "string",
                                "example": "/"
                              },
                              "subMenu": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "additionalProperties": true
                                }
                              }
                            }
                          }
                        },
                        "footer": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "example": "footer-menu-1"
                              },
                              "title": {
                                "type": "string",
                                "example": "Quick Links"
                              },
                              "links": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "additionalProperties": true
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Navigation configuration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server Error - Failed to fetch navigation data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/navigation/{id}": {
      "get": {
        "tags": [
          "Navigation"
        ],
        "summary": "Get specific navigation menu by ID",
        "description": "Returns a specific navigation menu structure by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the navigation menu to retrieve (header, footer, or a specific menu ID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved navigation menu",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "menu": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "header"
                        },
                        "name": {
                          "type": "string",
                          "example": "Header Navigation"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "title": {
                                "type": "string",
                                "example": "Home"
                              },
                              "href": {
                                "type": "string",
                                "example": "/"
                              },
                              "subMenu": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "additionalProperties": true
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Navigation menu not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server Error - Failed to fetch navigation menu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}